app not broken, updating look+feel
since removing thymeleaf, the app - when run via boot dashboard - isn't
creating the proper mappings to the groovy markupTemplateEngine inorder
to process .tpl files.  However, when run via gradle's bootRun task, the
app works fine.  

I changed the dropdown and single submit button to generate a random
item from an engram to a row of buttons for the options.  I created a
centered css control to center the buttons in the container and updated
the javascript to handle the multiple buttons.